Move initializer code in its own file
