+++ src/Nethermind/Nethermind.Blockchain/ReadOnlyBlockTree.cs
@@ -122,8 +122,6 @@ namespace Nethermind.Blockchain
 
         public bool WasProcessed(long number, Keccak blockHash) => _wrapped.WasProcessed(number, blockHash);
 
-        public void LoadLowestInsertedBeaconHeader() => _wrapped.LoadLowestInsertedBeaconHeader();
-
         public event EventHandler<BlockEventArgs> NewBestSuggestedBlock
         {
             add { }
